requirements.txt
streamlit
langchain
langchain-groq
python-dotenv
duckduckgo-search
pypdf